+++ packages/yourloops/test/integration/pages/patient-dashboard/patient-dashboard.caregiver.spec.tsx
@@ -35,9 +35,9 @@ import { mockUserApi } from '../../mock/user.api.mock'
 import { mockPatientApiForCaregivers } from '../../mock/patient.api.mock'
 import { UserRole } from '../../../../lib/auth/models/enums/user-role.enum'
 import { completeDashboardData, mockDataAPI } from '../../mock/data.api.mock'
-import { type AppMainLayoutParams, testAppMainLayoutForCaregiver } from '../../use-cases/app-main-layout-vizualization'
+import { type AppMainLayoutParams, testAppMainLayoutForCaregiver } from '../../use-cases/app-main-layout-visualisation'
 import { type PatientDashboardLayoutParams } from '../../assert/layout'
-import { testDashboardDataVisualisationPrivateTeam } from '../../use-cases/patient-data-vizualization'
+import { testDashboardDataVisualisationPrivateTeam } from '../../use-cases/patient-data-visualisation'
 
 describe('Patient dashboard for caregiver', () => {
   const monitoredPatientDashboardRoute = `/patient/${monitoredPatientId}/dashboard`
